Finish header design, fix dropdown issue
