+++ client/package.json
@@ -10,6 +10,7 @@
     "@zeit-ui/react-icons": "^1.2.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-headroom": "^3.0.0",
     "react-scripts": "^3.4.0"
   },
   "scripts": {
