+++ client/src/components/Header/Header.css
@@ -69,32 +69,6 @@
     }
 }
 
-.navbar .navbar-header .navbar-toggler {
-    border-radius: 5px;
-    background-color: transparent;
-    cursor: pointer;
-    border: none;
-    display: none;
-    outline: none;
-}
-
-@media (max-width: 768px) {
-    .navbar .navbar-header .navbar-toggler {
-        display: block;
-    }
-}
-
-.navbar .navbar-header .navbar-toggler span {
-    height: 2px;
-    width: 22px;
-    background-color: #929aad;
-    display: block;
-}
-
-.navbar .navbar-header .navbar-toggler span:not(:last-child) {
-    margin-bottom: 0.2rem;
-}
-
 .navbar .navbar-header>a {
     font-weight: 500;
     color: #3c4250;
@@ -109,7 +83,7 @@
 
 @media (max-width: 768px) {
     .navbar .navbar-menu {
-        display: none;
+        display: none;  /*Hides right nav on smaller screens*/
         text-align: center;
     }
 }
@@ -119,7 +93,7 @@
     flex-direction: row;
     display: flex;
     padding-left: 0;
-    margin-bottom: 0;
+    margin-bottom: 0.625rem;
     list-style: none;
 }
 
@@ -132,82 +106,22 @@
     }
 }
 
-.navbar .navbar-menu .navbar-nav>li>a {
-    color: #3c4250;
-    text-decoration: none;
-    display: inline-block;
-    padding: 0.5rem 1rem;
-}
-
-.navbar .navbar-menu .navbar-nav>li>a:hover {
-    color: #66f;
-}
-
-@media (max-width: 768px) {
-    .navbar .navbar-menu .navbar-nav>li>a {
-        border-bottom: 1px solid #EEE;
-    }
-}
-
-.navbar .navbar-menu .navbar-nav>li.active a {
-    color: #66f;
-}
-
-.navbar .navbar-menu .navbar-nav .navbar-dropdown .dropdown {
-    list-style: none;
-    position: absolute;
-    top: 150%;
-    left: 0;
-    background-color: #fff;
-    padding: 0.5rem 0;
-    min-width: 160px;
-    width: auto;
-    white-space: nowrap;
-    box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.1);
-    z-index: 99999;
-    border-radius: 0.75rem;
-    display: none;
-}
-
-@media (max-width: 768px) {
-    .navbar .navbar-menu .navbar-nav .navbar-dropdown .dropdown {
-        position: relative;
-        box-shadow: none;
-    }
-}
-
-.navbar .navbar-menu .navbar-nav .navbar-dropdown .dropdown li a {
-    color: #3c4250;
-    padding: 0.25rem 1rem;
-    display: block;
-}
-
-.navbar .navbar-menu .navbar-nav .navbar-dropdown .dropdown.show {
-    display: block !important;
-}
-
-.navbar .navbar-menu .navbar-nav .dropdown>.separator {
-    height: 1px;
-    width: 100%;
-    margin: 9px 0;
-    background-color: #eee;
-}
-
-.navbar .navbar-dropdown {
-    position: relative;
-}
-
 .navbar .navbar-header>a span {
-    color: #66f;
+    color: rgb(56, 56, 253);
 }
 
 .navbar .navbar-header h4 {
     font-weight: 500;
-    font-size: 1.25rem;
+    font-size: 2.25rem;
 }
 
 @media (max-width: 768px) {
     .navbar .navbar-header h4 {
         font-size: 1.05rem;
     }
+}
+
+.avatar{
+    width: 2.5rem !important;
+    height: 2.5rem !important;
 }
