+++ client/src/components/Header/navbarlogic.js
@@ -1,0 +1,56 @@
+let dropdowns = document.querySelectorAll('.navbar .dropdown-toggler');
+  let dropdownIsOpen = false;
+
+  // Handle dropdown menues
+  if (dropdowns.length) {
+    dropdowns.forEach((dropdown) => {
+      dropdown.addEventListener('click', (event) => {
+        let target = document.querySelector('#' + event.target.dataset.dropdown)
+
+        if (target) {
+          if (target.classList.contains('show')) {
+            target.classList.remove('show');
+            dropdownIsOpen = false;
+          } else {
+            target.classList.add('show');
+            dropdownIsOpen = true;
+          }
+        }
+      })
+    })
+  }
+
+  // Handle closing dropdowns if a user clicked the body
+  window.addEventListener('mouseup', (event) => {
+    if (dropdownIsOpen) {
+      dropdowns.forEach((dropdownButton) => {
+        let dropdown = document.querySelector('#' + dropdownButton.dataset.dropdown);
+        let targetIsDropdown = dropdown == event.target;
+
+        if (dropdownButton == event.target) {
+          return;
+        }
+
+        if ((!targetIsDropdown) && (!dropdown.contains(event.target))) {
+          dropdown.classList.remove('show');
+        }
+      });
+    }
+  });
+
+  // Open links in mobiles
+  function handleSmallScreens() {
+    document.querySelector('.navbar-toggler')
+      .addEventListener('click', () => {
+      let navbarMenu = document.querySelector('.navbar-menu')
+
+      if (navbarMenu.style.display === 'flex') {
+        navbarMenu.style.display = 'none';
+        return;
+      }
+
+      navbarMenu.style.display = 'flex';
+    });
+  }
+
+  handleSmallScreens();
